+++ src/interface/src/app/maplibre-map/map-nav-bar/map-nav-bar.component.html
@@ -1,0 +1,1 @@
+<ng-content></ng-content>
