+++ src/interface/src/app/maplibre-map/map-nav-bar/map-nav-bar.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { AsyncPipe, NgIf } from '@angular/common';
+import { BehaviorSubject } from 'rxjs';
+
+@Component({
+  selector: 'app-map-nav-bar',
+  standalone: true,
+  imports: [AsyncPipe, NgIf],
+  templateUrl: './map-nav-bar.component.html',
+  styleUrl: './map-nav-bar.component.scss',
+})
+export class MapNavbarComponent {
+  constructor() {}
+  opacity$ = new BehaviorSubject(0);
+}
