--- src/interface/src/app/maplibre-map/map-navbar/map-navbar.component.html
@@ -1,11 +1,0 @@
-<div class="map-navbar">
-  <app-map-base-dropdown></app-map-base-dropdown>
-  <sg-opacity-slider
-    class="navbar-slider"
-    *ngIf="true"
-    [minValue]="0"
-    [maxValue]="1"
-    [step]="0.01"
-    [sliderValue]="(opacity$ | async) || 0"
-    (valueChange)="handleOpacityChange($event)"></sg-opacity-slider>
-</div>
