+++ src/interface/src/app/maplibre-map/scenario-map/scenario-map.component.html
@@ -1,4 +1,13 @@
-<app-map-navbar></app-map-navbar>
+<app-map-nav-bar>
+  <app-map-base-dropdown></app-map-base-dropdown>
+  <sg-opacity-slider
+    class="navbar-slider"
+    [minValue]="0"
+    [maxValue]="1"
+    [step]="0.01"
+    [sliderValue]="(projectLayerOpacity$ | async) || 0"
+    (valueChange)="handleOpacityChange($event)"></sg-opacity-slider>
+</app-map-nav-bar>
 
 <mgl-map
   [style]="(baseLayerUrl$ | async) || ''"
