+++ src/interface/src/app/maplibre-map/scenario-map/scenario-map.component.scss
@@ -9,3 +9,9 @@
 mgl-map {
   flex: 1;
 }
+
+.navbar-slider {
+  ::ng-deep.slider-card {
+    height: 40px;
+  }
+}
