+++ src/interface/src/app/treatments/treatment-map/treatment-map.component.html
@@ -1,4 +1,4 @@
-<div class="map-navbar" *appFeatureFlag="'statewide_datalayers'">
+<app-map-nav-bar *appFeatureFlag="'statewide_datalayers'">
   <app-map-base-dropdown></app-map-base-dropdown>
   <sg-opacity-slider
     class="navbar-slider"
@@ -12,7 +12,7 @@
     [step]="0.01"
     [sliderValue]="(opacity$ | async) || 0"
     (valueChange)="handleOpacityChange($event)"></sg-opacity-slider>
-</div>
+</app-map-nav-bar>
 
 <mgl-map
   *ngIf="mapExtent$ | async; let bounds"
